requirements.txt
bokeh==2.3.3
gemmi==0.4.9
matplotlib==3.4.2
mendeleev==0.9.0
networkx==2.6.3
notebook==6.2.0
numpy==1.21.2
pandas==1.2.3
pymatgen==2020.10.20
torch==1.9.1
torch-cluster==1.5.9
torch-scatter==2.0.8
torch-sparse==0.6.12
torch-spline-conv==1.2.1
torch-geometric==2.0.1
scikit-learn==0.24.2
scipy==1.6.1
seaborn==0.11.2
selenium-firefox==2.0.7
torchaudio==0.9.1
torchvision==0.2.2
torchviz==0.0.2
